refactor(admin): split action persistence helpers
Move low-level D1 page-content saves and guarded thought/atom row helpers out of root admin actions so actions.ts stays focused on exported server actions.

Verified with pnpm --filter @anipotts/admin typecheck, pnpm --filter @anipotts/admin lint, pnpm --filter @anipotts/admin build, pnpm exec prettier --check apps/admin/src/app/actions.ts apps/admin/src/app/lib/content-records.ts apps/admin/src/app/lib/page-content.ts, git diff --check -- apps/admin/src/app/actions.ts apps/admin/src/app/lib/content-records.ts apps/admin/src/app/lib/page-content.ts, and pnpm dlx knip --no-exit-code.
